tmf: Ellipse function names that don't fit in CallStackView
Add "..." to function names that get cut due to short rectangles
in the call stack view, and avoid printing anything if at least
1 character + the "..." won't fit.

Also, instead of starting from the full string length and going
down to find a fitting string, start from 1 and work our way
up. This will better handle the worst case, which is lots of
very small states in a view that all want to print very long
strings.

Bug: 497111

Change-Id: I94ccca31fb890923063f75359ff714721a8b6f68
Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>