tmf: Add Export Time Selection action and related interface
Traces implementing the new ITmfTrimmableTrace interface should provide
the "Export Time Selection" action, which will prompt the user for a
directory, and trim the trace (using the current selected time range) and
save the new cut trace in said directory.

The implementation of the trim operation itself is trace-type specific.

Change-Id: I02440b49dc7b84465f4ce02ecacc2604bbb90946
Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>