os.linux & tmf.ui: introduce dynamic filter for cfv: Active threads and Threads on CPUs
A dynamic filter is defined as a filter that is calculated on each
update of the visual time range.

Users can now apply dynamic filters on the Control Flow view. A new
sub menu of the view menu allows users to enable and configure those
dynamic filters. For now, one dynamic filter is available: "Active
Threads Only". This filter has two configurable options:

    - Show all threads considered active.
    - Show threads running on a selection of CPUs.

A thread is considered inactive if its state is one of the following:

    - non-existing
    - unknown
    - wait (blocked)
    - wait (unknown)

Otherwise the thread is considered active.

The "Show threads running on a selection of CPUs" filter shows
threads that were currently running or waiting on the run queue for the
ranges of CPUs supplied to the filter during the visible time range of
the Control Flow view.

Change-Id: Ib58fe09aabd070e4db9270955f86737c36f309ca
Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>