ss: Add quark dimension to narrow down search queries and rsht
The State History Tree was considered to only have the time
dimension. However most queries ask for specific quarks too.
We can benefit from using the range of quarks to further
narrow down the nodes which need to be searched in the sht.

This patch adds the quark bounds to node headers and adapts
relevant methods. This implements a new statesystem-specific
tree dubbed the Enhanced History Tree (eSHT). Note this patch
only adds this implementation and relevant tests, it does not
make HistoryTreeBackend (and the "real" state system) make
use of it yet.

TODO
- Implement the eSHT-specific behaviour in EnhancedNode.

Change-Id: Id029e4fe56e946f40f1d05a25b6c398e00863007
Signed-off-by: Loïc Prieur-Drevon <loic.prieurdrevon@gmail.com>
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>