tmf: Rework how analysis module help text is handled
The notions of getFullText() and getShortText(ITmfTrace) was
not very clear. Rename those to

getGenericHelpText()
getTraceSpecificHelpText(ITmfTrace)

and use templating in TmfAbstractAnalysisModule to ensure
sub-classes override these methods, and not the interface
getHelpText() ones.

With the new approach, analyses now define the generic
and trace-specific messages separately, in a way that the
generic text is always shown (even when a trace is opened)
without having to repeat the same information in both methods.

This incurs API breaks, so will have to wait for the 3.0 branch
to open.

Change-Id: I5b4da030b456ccfc13beae397e9a575d551052b1
Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>