tmf: Add the notion of required event names to analyses
Add a new protected method to allow analysis modules to list
the names of events (or others) that are required for this
analysis.

It might be tempting to put this information in the trace-
specific text instead, and make use of the event layout to
get the real event names, but remember: this information
is for the user who is *about to* take a trace, and wants
to know which events he needs to enable. It is not for traces
that already exist.

This could be integrated with the notion of analysis
requirements someday, but most analyses do not specify those
requirements, so I propose a flat list of strings for now,
so that at least the user has something to work with.

Change-Id: Ib0ee8f759ad6ef312b308cf0a886e7116f7cff14
Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>